apps/servicebus-processor/requirements.txt
Flask
azure-servicebus

